RavenDB-22362 address PR comments
