+++ src/Raven.Server/Web/RequestHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.IO.Compression;
@@ -26,11 +25,10 @@ using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.TrafficWatch;
 using Sparrow;
-using Sparrow.Logging;
 
 namespace Raven.Server.Web
 {
-    public abstract class RequestHandler
+    public abstract partial class RequestHandler
     {
         public const string StartParameter = "start";
 
@@ -763,45 +761,6 @@ namespace Raven.Server.Web
         public void AddStringToHttpContext(string str, TrafficWatchChangeType type)
         {
             HttpContext.Items["TrafficWatch"] = (str, type);
-        }
-
-        public bool IsLocalRequest()
-        {
-            if (HttpContext.Connection.RemoteIpAddress == null && HttpContext.Connection.LocalIpAddress == null)
-            {
-                return true;
-            }
-            if (HttpContext.Connection.RemoteIpAddress.Equals(HttpContext.Connection.LocalIpAddress))
-            {
-                return true;
-            }
-            if (IPAddress.IsLoopback(HttpContext.Connection.RemoteIpAddress))
-            {
-                return true;
-            }
-            return false;
-        }
-
-        public string RequestIp => IsLocalRequest() ? Environment.MachineName : HttpContext.Connection.RemoteIpAddress.ToString();
-
-        public void LogAuditFor(string logger, string message)
-        {
-            var auditLog = LoggingSource.AuditLog.GetLogger(logger, "Audit");
-            Debug.Assert(auditLog.IsInfoEnabled, $"auditlog info is disabled");
-
-            var clientCert = GetCurrentCertificate();
-            
-            var sb = new StringBuilder();
-            sb.Append(RequestIp);
-            sb.Append(", ");
-            if (clientCert != null) 
-                sb.Append($"CN={clientCert.Subject} [{clientCert.Thumbprint}], ");
-            else
-                sb.Append("no certificate, ");
-            
-            sb.Append(message);
-
-            auditLog.Info(sb.ToString());
         }
     }
 }
